backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.0
chromadb==0.4.14
pypdf==3.17.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sentence-transformers==5.1.2
